refactor: remove StartGraphCmd interface and related functions from Go command implementation
