+++ core/src/ten_runtime/binding/go/interface/ten_runtime/cmd.go
@@ -26,15 +26,6 @@ type Cmd interface {
 	Clone() (Cmd, error)
 }
 
-// StartGraphCmd is the interface for the start graph command.
-type StartGraphCmd interface {
-	Cmd
-
-	SetPredefinedGraphName(predefinedGraphName string) error
-	SetGraphFromJSONBytes(graphJSONBytes []byte) error
-	SetLongRunningMode(longRunningMode bool) error
-}
-
 // NewCmd creates a custom cmd which is intended to be sent to another
 // extension using tenEnv.SendCmd().
 func NewCmd(cmdName string) (Cmd, error) {
@@ -115,24 +106,6 @@ func NewCmd(cmdName string) (Cmd, error) {
 	return newCmd(bridge), nil
 }
 
-// NewStartGraphCmd creates a new start graph command.
-func NewStartGraphCmd() (StartGraphCmd, error) {
-	var bridge C.uintptr_t
-	err := withCGOLimiter(func() error {
-		cStatus := C.ten_go_cmd_create_start_graph_cmd(
-			&bridge,
-		)
-		e := withCGoError(&cStatus)
-
-		return e
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return newStartGraphCmd(bridge), nil
-}
-
 type cmd struct {
 	*msg
 }
@@ -140,16 +113,6 @@ type cmd struct {
 func newCmd(bridge C.uintptr_t) *cmd {
 	return &cmd{
 		msg: newMsg(bridge),
-	}
-}
-
-type startGraphCmd struct {
-	*cmd
-}
-
-func newStartGraphCmd(bridge C.uintptr_t) *startGraphCmd {
-	return &startGraphCmd{
-		cmd: newCmd(bridge),
 	}
 }
 
@@ -178,51 +141,4 @@ func (p *cmd) Clone() (Cmd, error) {
 	return newCmd(bridge), nil
 }
 
-func (p *startGraphCmd) SetPredefinedGraphName(
-	predefinedGraphName string,
-) error {
-	defer p.keepAlive()
-
-	err := withCGOLimiter(func() error {
-		apiStatus := C.ten_go_cmd_start_graph_set_predefined_graph_name(
-			p.getCPtr(),
-			unsafe.Pointer(unsafe.StringData(predefinedGraphName)),
-			C.int(len(predefinedGraphName)),
-		)
-		return withCGoError(&apiStatus)
-	})
-
-	return err
-}
-
-func (p *startGraphCmd) SetGraphFromJSONBytes(graphJSONBytes []byte) error {
-	defer p.keepAlive()
-
-	err := withCGOLimiter(func() error {
-		apiStatus := C.ten_go_cmd_start_graph_set_graph_from_json_bytes(
-			p.getCPtr(),
-			unsafe.Pointer(unsafe.SliceData(graphJSONBytes)),
-			C.int(len(graphJSONBytes)),
-		)
-		return withCGoError(&apiStatus)
-	})
-
-	return err
-}
-
-func (p *startGraphCmd) SetLongRunningMode(longRunningMode bool) error {
-	defer p.keepAlive()
-
-	err := withCGOLimiter(func() error {
-		apiStatus := C.ten_go_cmd_start_graph_set_long_running_mode(
-			p.getCPtr(),
-			C.bool(longRunningMode),
-		)
-		return withCGoError(&apiStatus)
-	})
-
-	return err
-}
-
 var _ Cmd = new(cmd)
-var _ StartGraphCmd = new(startGraphCmd)
